refactor: move visibility control under ros2/
