+++ include/kalman_filter_localization/ros2/visibility_control.hpp
@@ -29,12 +29,37 @@
 // LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN
 // ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
 // POSSIBILITY OF SUCH DAMAGE.
-#ifndef KALMAN_FILTER_LOCALIZATION__VISIBILITY_CONTROL_HPP_
-#define KALMAN_FILTER_LOCALIZATION__VISIBILITY_CONTROL_HPP_
+#ifndef KALMAN_FILTER_LOCALIZATION__ROS2__VISIBILITY_CONTROL_HPP_
+#define KALMAN_FILTER_LOCALIZATION__ROS2__VISIBILITY_CONTROL_HPP_
 
-// Compatibility header:
-// Keep the original include path while moving ROS2-specific headers under
-// include/kalman_filter_localization/ros2.
-#include <kalman_filter_localization/ros2/visibility_control.hpp>
+// The below macros are taken from https://gcc.gnu.org/wiki/Visibility and from
+// demos/composition/include/composition/visibility_control.h at https://github.com/ros2/demos
+#if defined _WIN32 || defined __CYGWIN__
+  #ifdef __GNUC__
+    #define KFL_EKFL_EXPORT __attribute__ ((dllexport))
+    #define KFL_EKFL_IMPORT __attribute__ ((dllimport))
+  #else
+    #define KFL_EKFL_EXPORT __declspec(dllexport)
+    #define KFL_EKFL_IMPORT __declspec(dllimport)
+  #endif
+  #ifdef KFL_EKFL_BUILDING_DLL
+    #define KFL_EKFL_PUBLIC KFL_EKFL_EXPORT
+  #else
+    #define KFL_EKFL_PUBLIC KFL_EKFL_IMPORT
+  #endif
+  #define KFL_EKFL_PUBLIC_TYPE KFL_EKFL_PUBLIC
+  #define KFL_EKFL_LOCAL
+#else
+  #define KFL_EKFL_EXPORT __attribute__ ((visibility("default")))
+  #define KFL_EKFL_IMPORT
+  #if __GNUC__ >= 4
+    #define KFL_EKFL_PUBLIC __attribute__ ((visibility("default")))
+    #define KFL_EKFL_LOCAL  __attribute__ ((visibility("hidden")))
+  #else
+    #define KFL_EKFL_PUBLIC
+    #define KFL_EKFL_LOCAL
+  #endif
+  #define KFL_EKFL_PUBLIC_TYPE
+#endif
 
-#endif  // KALMAN_FILTER_LOCALIZATION__VISIBILITY_CONTROL_HPP_
+#endif  // KALMAN_FILTER_LOCALIZATION__ROS2__VISIBILITY_CONTROL_HPP_
